Order Works? Yeah?
OK order-ahm-'works' by displaying menu items then user gets to choose menu item and size yay. not done by a chatbot because idk what to do for that actually. so might as well make systems around that

anyhow, the toppings will come next and is placed between both addthensize and final adding which will display certain toppings depending on what the order selected

no one but me will read this. this is a note for me to remember,